# docs/api/api_docs.yaml
openapi: 3.0.3
info:
  title: CAB System API Gateway
  version: 1.0.0
  description: |
    REST contract exposed by the CAB System API Gateway.
    Client/Postman calls port 3000. The Gateway validates JWT and calls
    microservices through gRPC. Internal service HTTP ports are not part of the
    client contract.
servers:
  - url: http://localhost:3000
    description: Local Docker Compose Gateway
security:
  - bearerAuth: []
tags:
  - name: Health
  - name: Auth
  - name: Customer
  - name: Driver
  - name: Booking
  - name: Trip
  - name: Payment
  - name: Notification
paths:
  /health:
    get:
      tags: [Health]
      security: []
      responses:
        '200':
          description: Gateway process is running
  /ready:
    get:
      tags: [Health]
      security: []
      responses:
        '200':
          description: Gateway and dependencies are ready
        '503':
          description: One or more dependencies are unavailable
  /health/services:
    get:
      tags: [Health]
      security: []
      responses:
        '200':
          description: All services are healthy
        '503':
          description: At least one service is down
  /api/:
    get:
      tags: [Health]
      security: []
      responses:
        '200':
          description: API Gateway information
  /api/auth/register:
    post:
      tags: [Auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201': { description: Account created }
        '400': { description: Invalid request }
        '409': { description: Account already exists }
  /api/auth/login:
    post:
      tags: [Auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200': { description: Login successful }
        '401': { description: Invalid credentials }
  /api/auth/driver/send-otp:
    post:
      tags: [Auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone]
              properties:
                phone: { type: string }
      responses:
        '200': { description: OTP sent }
  /api/auth/driver/verify-otp:
    post:
      tags: [Auth]
      security: []
      responses:
        '200': { description: OTP verified }
        '400': { description: Invalid OTP }
  /api/auth/driver/register:
    post:
      tags: [Auth]
      security: []
      responses:
        '201': { description: Driver application created }
        '400': { description: Invalid application }
  /api/auth/driver/set-password:
    post:
      tags: [Auth]
      security: []
      responses:
        '200': { description: Driver password set }
  /api/customers/{userId}:
    get:
      tags: [Customer]
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200': { description: Customer returned }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    put:
      tags: [Customer]
      parameters:
        - $ref: '#/components/parameters/UserId'
      requestBody:
        required: true
        content:
          application/json: { schema: { type: object } }
      responses:
        '200': { description: Customer updated }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/drivers:
    get:
      tags: [Driver]
      responses:
        '200': { description: Driver list }
  /api/drivers/{id}:
    get:
      tags: [Driver]
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200': { description: Driver returned }
    put:
      tags: [Driver]
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json: { schema: { type: object } }
      responses:
        '200': { description: Driver updated }
  /api/drivers/nearby:
    get:
      tags: [Driver]
      description: Returns online drivers within the fixed default radius of 1 km.
      parameters:
        - { name: lat, in: query, required: true, schema: { type: number } }
        - { name: lng, in: query, required: true, schema: { type: number } }
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: limit, in: query, schema: { type: integer, default: 20 } }
      responses:
        '200': { description: Nearby driver list with pagination }
        '400': { description: Invalid coordinates }
  /api/drivers/online:
    get:
      tags: [Driver]
      description: Operator/Admin endpoint for online drivers.
      responses:
        '200': { description: Online driver list }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/drivers/{id}/location:
    get:
      tags: [Driver]
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200': { description: Driver location }
    put:
      tags: [Driver]
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json: { schema: { type: object } }
      responses:
        '200': { description: Location updated }
  /api/drivers/{id}/online:
    put:
      tags: [Driver]
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200': { description: Driver is online }
  /api/drivers/{id}/offline:
    put:
      tags: [Driver]
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200': { description: Driver is offline }
  /api/drivers/applications:
    get:
      tags: [Driver]
      responses:
        '200': { description: Application list }
  /api/drivers/applications/{id}/approve:
    put:
      tags: [Driver]
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200': { description: Application approved }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/drivers/applications/{id}/reject:
    put:
      tags: [Driver]
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200': { description: Application rejected }
  /api/drivers/applications/{id}/resubmit:
    put:
      tags: [Driver]
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200': { description: Application resubmitted }
  /api/bookings:
    post:
      tags: [Booking]
      description: Customer ID is derived from the JWT, not the request body.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookingCreate'
      responses:
        '201': { description: Booking created }
        '401': { $ref: '#/components/responses/Unauthorized' }
    get:
      tags: [Booking]
      parameters:
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: limit, in: query, schema: { type: integer, default: 20 } }
      responses:
        '200': { description: Booking list with pagination }
  /api/bookings/{id}:
    get:
      tags: [Booking]
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200': { description: Booking returned }
  /api/bookings/{id}/accept:
    put:
      tags: [Booking]
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200': { description: Booking accepted and Trip created }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/bookings/{id}/reject:
    put:
      tags: [Booking]
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200': { description: Booking rejected or reassigned }
  /api/bookings/{id}/cancel:
    put:
      tags: [Booking]
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        content:
          application/json: { schema: { type: object } }
      responses:
        '200': { description: Booking and related Trip canceled }
  /api/bookings/{id}/status:
    put:
      tags: [Booking]
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200': { description: Booking status updated }
  /api/trips:
    post:
      tags: [Trip]
      description: Operator/Admin endpoint for direct Trip creation.
      responses:
        '201': { description: Trip created }
    get:
      tags: [Trip]
      responses:
        '200': { description: Trip history }
  /api/trips/{id}:
    get:
      tags: [Trip]
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200': { description: Trip returned }
  /api/trips/{id}/status:
    put:
      tags: [Trip]
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json: { schema: { type: object } }
      responses:
        '200': { description: Trip status updated }
  /api/trips/{id}/rating:
    get:
      tags: [Trip]
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200': { description: Rating returned }
    post:
      tags: [Trip]
      description: Reviewer identity is derived from the JWT.
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json: { schema: { $ref: '#/components/schemas/Rating' } }
      responses:
        '201': { description: Rating submitted }
  /api/payments:
    post:
      tags: [Payment]
      description: Amount and customer identity are derived server-side.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '201': { description: Payment created }
    get:
      tags: [Payment]
      responses:
        '200': { description: Payment returned/listed }
  /api/payments/estimate/{tripId}:
    get:
      tags: [Payment]
      parameters:
        - $ref: '#/components/parameters/TripId'
      responses:
        '200': { description: Fare estimate }
  /api/payments/{id}:
    get:
      tags: [Payment]
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200': { description: Payment returned }
  /api/payments/{id}/process:
    put:
      tags: [Payment]
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200': { description: Payment processed }
  /api/notifications:
    post:
      tags: [Notification]
      responses:
        '201': { description: Notification created }
  /api/notifications/recipient/{recipientId}:
    get:
      tags: [Notification]
      parameters:
        - name: recipientId
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200': { description: Recipient notifications }
  /api/notifications/{id}:
    get:
      tags: [Notification]
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200': { description: Notification returned }
  /api/notifications/{id}/read:
    put:
      tags: [Notification]
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200': { description: Notification marked as read }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: integer }
    UserId:
      name: userId
      in: path
      required: true
      schema: { type: integer }
    TripId:
      name: tripId
      in: path
      required: true
      schema: { type: integer }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema: { type: string }
  responses:
    Unauthorized:
      description: Missing or invalid JWT
    Forbidden:
      description: Role or ownership check failed
  schemas:
    RegisterRequest:
      type: object
      required: [phone, password, role]
      properties:
        phone: { type: string }
        email: { type: string, format: email }
        password: { type: string, minLength: 8, format: password }
        role: { type: string, enum: [CUSTOMER, DRIVER, ADMIN, OPERATOR] }
        fullName: { type: string }
    LoginRequest:
      type: object
      required: [password]
      properties:
        identifier: { type: string }
        email: { type: string }
        phone: { type: string }
        password: { type: string, format: password }
    BookingCreate:
      type: object
      required: [pickLat, pickLng, destLat, destLng, vehicleType]
      properties:
        pickLat: { type: number }
        pickLng: { type: number }
        destLat: { type: number }
        destLng: { type: number }
        vehicleType: { type: string }
    Rating:
      type: object
      required: [rating]
      properties:
        rating: { type: integer, minimum: 1, maximum: 5 }
        comment: { type: string }